✨ add building charts
