+++ viewer/src/logic/charts.ts
@@ -1,23 +1,4 @@
-import Engine, {
-  AdvTechTilePos,
-  BoardAction,
-  Booster,
-  Command,
-  EventSource,
-  Faction,
-  factionBoard,
-  factions,
-  finalRankings,
-  gainFinalScoringVictoryPoints,
-  LogEntry,
-  Planet,
-  Player,
-  PlayerEnum,
-  ResearchField,
-  Resource,
-  RoundScoring,
-  TechPos,
-} from "@gaia-project/engine";
+import Engine, { EventSource, factions, LogEntry, Planet, Player, PlayerEnum, Resource } from "@gaia-project/engine";
 import {
   ChartConfiguration,
   ChartDataset,
@@ -28,15 +9,15 @@ import {
   TooltipModel,
   TooltipOptions,
 } from "chart.js";
-import { sortBy, sum, sumBy } from "lodash";
+import { sumBy } from "lodash";
 import { planetColor } from "../graphics/utils";
-import { parseCommands } from "./recent";
+import { CommandObject, parseCommands } from "./recent";
 
-type DeepPartial<T> = {
+export type DeepPartial<T> = {
   [P in keyof T]?: DeepPartial<T[P]>;
 };
 
-type EventFilter = (
+export type EventFilter = (
   player: PlayerEnum,
   source: EventSource,
   resource: Resource,
@@ -44,7 +25,7 @@ type EventFilter = (
   change: number
 ) => number;
 
-interface DatasetFactory {
+export interface DatasetFactory {
   backgroundColor: string;
   fill: string | false;
   label: string;
@@ -53,233 +34,17 @@ interface DatasetFactory {
   weight: number;
 }
 
-function simulateIncome(pl: Player, consume: (p: Player) => void): number {
-  const json = JSON.parse(JSON.stringify(pl));
-  delete json.federationCache; // otherwise we need to pass the map when loading
-  const clone = Player.fromData(json, null, 0);
-  consume(clone);
-  return clone.data.victoryPoints - pl.data.victoryPoints;
-}
-
-type VictoryPointAggregate = {
-  label: string;
-  description: string;
-  color: string;
-};
-
-const charge: VictoryPointAggregate = {
-  label: "Charge",
-  description: "10 starting points - initial bid - points spend for power charges",
-  color: "--res-power",
-};
-
-const finalScoring: VictoryPointAggregate = {
-  label: "Final",
-  description: "Final Scoring A + B",
-  color: "--rt-sci",
-};
-
-const otherScoring: VictoryPointAggregate = {
-  label: "Other",
-  description: "Base Tech Tiles, Gleens special",
-  color: "--tech-tile",
-};
-
-type VictoryPointType = EventSource | "chart-init" | "chart-bid" | "chart-spend" | "chart-final1" | "chart-final2";
-
-const aggregateColor = "black"; //this is never displayed, because it's only in the bar chart
-
-type IncludeRounds = "all" | "except-final" | "last";
-
-export type ResourceKind = Resource | "pay-pw";
-
-export type ResourceSource = {
-  type: ResourceKind;
-  inverseOf?: Resource;
-  label: string;
-  plural: string;
-  color: string;
-  weight: number;
-};
-
-export const resourceSources: ResourceSource[] = [
-  {
-    type: Resource.Credit,
-    label: "Credit",
-    plural: "Credits",
-    color: "--res-credit",
-    weight: 1,
-  },
-  {
-    type: Resource.Ore,
-    label: "Ore",
-    plural: "Ores",
-    color: "--res-ore",
-    weight: 3,
-  },
-  {
-    type: Resource.Knowledge,
-    label: "Knowledge",
-    plural: "Knowledge",
-    color: "--res-knowledge",
-    weight: 4,
-  },
-  {
-    type: Resource.Qic,
-    label: "QIC",
-    plural: "QICs",
-    color: "--res-qic",
-    weight: 4,
-  },
-  {
-    type: Resource.ChargePower,
-    label: "Power Charges",
-    plural: "Power Charges",
-    color: "--res-power",
-    weight: 0,
-  },
-  {
-    type: "pay-pw",
-    inverseOf: Resource.ChargePower,
-    label: "Spent Power",
-    plural: "Spent Power",
-    color: "--lost",
-    weight: 0,
-  },
-  {
-    type: Resource.GainToken,
-    label: "Gained Tokens",
-    plural: "Gained Tokens",
-    color: "--recent",
-    weight: 0,
-  },
-];
-
-type VictoryPointSource = {
-  type: VictoryPointType[];
-  label: string;
-  description: string;
-  color: string;
-  projectedEndValue?: (e: Engine, p: Player) => number;
-  initialValue?: (p: Player) => number;
-  aggregate?: VictoryPointAggregate;
-};
-
-const victoryPointSources: VictoryPointSource[] = [
-  {
-    type: ["chart-init"],
-    label: "Initial",
-    description: "10 starting points",
-    color: aggregateColor,
-    aggregate: charge,
-    initialValue: () => 10,
-  },
-  {
-    type: ["chart-bid"],
-    label: "Bid",
-    description: "Initial bid",
-    color: aggregateColor,
-    aggregate: charge,
-    initialValue: (p) => -p.data.bid,
-  },
-  {
-    type: [Command.ChargePower],
-    label: "Charge",
-    description: "Points spend for power charges",
-    color: aggregateColor,
-    aggregate: charge,
-  },
-  {
-    type: TechPos.values(),
-    label: "Base Tech",
-    description: "Base Tech Tiles",
-    color: aggregateColor,
-    aggregate: otherScoring,
-  },
-  {
-    type: [Faction.Gleens],
-    label: "Gleens",
-    description: "Gleens special",
-    color: aggregateColor,
-    aggregate: otherScoring,
-  },
-  {
-    type: RoundScoring.values(),
-    label: "Round",
-    description: "Round Bonus",
-    color: "--rt-gaia",
-  },
-  {
-    type: Booster.values(),
-    label: "Booster",
-    description: "Round Boosters",
-    color: "--oxide",
-    projectedEndValue: (e, p) => (e.ended ? 0 : simulateIncome(p, (clone) => clone.receivePassIncome())),
-  },
-  {
-    type: BoardAction.values(),
-    label: "QIC",
-    description: "QIC actions",
-    color: "--specialAction",
-  },
-  {
-    type: ([Command.UpgradeResearch] as VictoryPointType[]).concat(ResearchField.values()),
-    label: "Research",
-    description: "Research in counted in the round where the level is reached",
-    color: "--res-knowledge",
-  },
-  {
-    type: AdvTechTilePos.values(),
-    label: "Advanced Tech",
-    description: "Advanced Tech Tiles",
-    color: "--current-round",
-  },
-  {
-    type: [Command.FormFederation],
-    label: "Federation",
-    description: "Re-scoring is counted as QIC action",
-    color: "--federation",
-  },
-  {
-    type: ["chart-final1"],
-    label: "Final A",
-    description: "Final Scoring A",
-    color: aggregateColor,
-    aggregate: finalScoring,
-    projectedEndValue: (engine: Engine, pl: Player) =>
-      simulateIncome(pl, (clone) =>
-        gainFinalScoringVictoryPoints(finalRankings([engine.tiles.scorings.final[0]], engine.players), clone)
-      ),
-  },
-  {
-    type: ["chart-final2"],
-    label: "Final B",
-    description: "Final Scoring B",
-    color: aggregateColor,
-    aggregate: finalScoring,
-    projectedEndValue: (engine: Engine, pl: Player) =>
-      simulateIncome(pl, (clone) =>
-        gainFinalScoringVictoryPoints(finalRankings([engine.tiles.scorings.final[1]], engine.players), clone)
-      ),
-  },
-  {
-    type: ["chart-spend"],
-    label: "Resources",
-    description: "Points for the remaining resources converted to credits",
-    color: "--res-credit",
-    projectedEndValue: (_: Engine, pl: Player) => simulateIncome(pl, (clone) => clone.data.finalResourceHandling()),
-  },
-];
+export type IncludeRounds = "all" | "except-final" | "last";
 
 export function chartPlayerOrder(data: Engine) {
   return data.players.map((p) => p.player);
 }
 
-function getDataPoints(
+export function getDataPoints(
   data: Engine,
   initialValue: number,
   extractChange: EventFilter,
-  extractLog: (LogEntry) => number,
+  extractLog: (moveHistory: string[], entry: LogEntry) => number,
   projectedEndValue: () => number,
   deltaForEnded: () => number,
   includeRounds: IncludeRounds
@@ -298,7 +63,7 @@ function getDataPoints(
       round = logItem.round;
     }
 
-    counter += extractLog(logItem);
+    counter += extractLog(data.moveHistory, logItem);
 
     const changes = logItem.changes;
     if (changes) {
@@ -332,175 +97,16 @@ function getDataPoints(
   return perRoundData;
 }
 
-export function countResearch(moveHistory: string[], faction: Faction) {
-  const research = new Map<ResearchField, number>();
-  factionBoard(faction).income[0].rewards.forEach((r) => {
-    if (r.type.startsWith("up-")) {
-      research.set(r.type.slice(3) as ResearchField, 1);
-    }
-  });
-
-  return (log: LogEntry) => {
-    let delta = 0;
-    if (log.move != null) {
-      const move = moveHistory[log.move]; // current move isn't added yet
-      if (move != null) {
-        for (const cmd of parseCommands(move)) {
-          if (cmd.faction == faction && cmd.command == Command.UpgradeResearch) {
-            const field = cmd.args[0] as ResearchField;
-            const newLevel = (research.get(field) ?? 0) + 1;
-            research.set(field, newLevel);
-            if (newLevel >= 3) {
-              delta += 4;
-            }
-          }
-        }
-      }
-    }
-    return delta;
-  };
-}
-
-function groupSources(sources: VictoryPointSource[]): VictoryPointSource[] {
-  const res: VictoryPointSource[] = [];
-
-  const groupTypes = new Map<string, VictoryPointSource>();
-
-  for (const source of sources) {
-    const aggregate = source.aggregate;
-    if (aggregate != null) {
-      let group = groupTypes.get(aggregate.label);
-      if (group == null) {
-        group = {
-          type: [],
-          label: aggregate.label,
-          description: aggregate.description,
-          color: aggregate.color,
-          projectedEndValue: (e, p) => 0,
-          initialValue: (p) => 0,
-        };
-        groupTypes.set(aggregate.label, group);
-        res.push(group);
-      }
-      group.type.push(...source.type);
-      if (source.projectedEndValue != null) {
-        const last = group.projectedEndValue;
-        group.projectedEndValue = (e, p) => last(e, p) + source.projectedEndValue(e, p);
-      }
-      if (source.initialValue != null) {
-        const last = group.initialValue;
-        group.initialValue = (p) => last(p) + source.initialValue(p);
-      }
-    } else {
-      res.push(source);
-    }
-  }
-
-  return res;
-}
-
-function victoryPointDetails(
-  data: Engine,
-  player: PlayerEnum,
-  canvas: HTMLCanvasElement,
-  sources: VictoryPointSource[]
-): DatasetFactory[] {
-  const pl = data.player(player);
-  if (!pl.faction) {
-    return [];
-  }
-  const wantPlayer = player;
-
-  const style = window.getComputedStyle(canvas);
-
-  return sources.map((s) => {
-    const type = s.type;
-    const values = Object.values(type) as VictoryPointType[];
-    const color = style.getPropertyValue(s.color);
-
-    const initialValue = s.initialValue != null ? s.initialValue(pl) : 0;
-
-    const extractChange = (player, source, resource, round, change) =>
-      player == wantPlayer && resource == Resource.VictoryPoint && values.includes(source) ? change : 0;
-
-    const extractLog = s.type.includes(Command.UpgradeResearch) ? countResearch(data.moveHistory, pl.faction) : () => 0;
-
-    const deltaForEnded = () => {
-      if (s.type.includes(Command.UpgradeResearch)) {
-        // research was already counted in chart separately
-        return -simulateIncome(pl, (clone) => clone.data.gainResearchVictoryPoints());
-      }
-      return 0;
-    };
-
-    return {
-      backgroundColor: color,
-      label: s.label,
-      description: s.description,
-      fill: "-1",
-      getDataPoints: () =>
-        getDataPoints(
-          data,
-          initialValue,
-          extractChange,
-          extractLog,
-          () => (s.projectedEndValue == null ? 0 : s.projectedEndValue(data, pl)),
-          deltaForEnded,
-          "all"
-        ),
-      weight: 1,
-    };
-  });
-}
-
-function resourceDetails(
-  data: Engine,
-  player: PlayerEnum,
-  canvas: HTMLCanvasElement,
-  sources: ResourceSource[],
-  includeRounds: IncludeRounds
-): DatasetFactory[] {
-  const pl = data.player(player);
-  if (!pl.faction) {
-    return [];
-  }
-  const wantPlayer = player;
-
-  const style = window.getComputedStyle(canvas);
-
-  return sources.map((s) => {
-    const color = style.getPropertyValue(s.color);
-
-    const initialValue = 0;
-    const extractChange = (player, source, resource, round, change) =>
-      player == wantPlayer && ((resource == s.type && change > 0) || (resource == s.inverseOf && change < 0))
-        ? Math.abs(change)
-        : 0;
-
-    const extractLog = () => 0;
-    const deltaForEnded = () => 0;
-
-    return {
-      backgroundColor: color,
-      label: s.label,
-      fill: false,
-      getDataPoints: () =>
-        getDataPoints(data, initialValue, extractChange, extractLog, () => 0, deltaForEnded, includeRounds),
-      weight: s.weight,
-    };
-  });
-}
-
-function playerColor(pl: Player, invert: boolean) {
+export function playerColor(pl: Player, invert: boolean) {
   const planet: Planet = factions[pl.faction].planet;
   return invert && planet == Planet.Ice ? "black" : planetColor(planet);
 }
 
-function playerLabel(pl: Player) {
+export function playerLabel(pl: Player) {
   return factions[pl.faction].name;
 }
 
-function weightedSum(
+export function weightedSum(
   data: Engine,
   player: PlayerEnum,
   canvas: HTMLCanvasElement,
@@ -528,6 +134,28 @@ function weightedSum(
   };
 }
 
+export function logEntryProcessor(
+  player: Player,
+  processor: (cmd: CommandObject) => number
+): (moveHistory: string[], log: LogEntry) => number {
+  return (moveHistory: string[], log: LogEntry): number => {
+    let res = 0;
+
+    if (log.move != null) {
+      const move = moveHistory[log.move]; // current move isn't added yet
+      if (move != null) {
+        for (const cmd of parseCommands(move)) {
+          if (cmd.faction == player.faction) {
+            res += processor(cmd);
+          }
+        }
+      }
+    }
+
+    return res;
+  };
+}
+
 function newLegendOptions(provider: (index: number) => string) {
   let hovering = false;
   const tooltip = document.getElementById("tooltip");
@@ -551,7 +179,7 @@ function newLegendOptions(provider: (index: number) => string) {
   return legendOptions;
 }
 
-function lineChartConfig(
+export function lineChartConfig(
   title: string,
   canvas: HTMLCanvasElement,
   data: Engine,
@@ -615,171 +243,4 @@ function lineChartConfig(
       },
     },
   };
-}
-
-export function newVictoryPointsBarChart(data: Engine, canvas: HTMLCanvasElement): ChartConfiguration<"bar"> {
-  const datasets: ChartDataset<"bar">[] = sortBy(
-    chartPlayerOrder(data)
-      .filter((pl) => data.player(pl).faction != null)
-      .map((pl) => {
-        const points = victoryPointDetails(data, pl, canvas, victoryPointSources).map((f) => f.getDataPoints()[7]);
-        const player = data.player(pl);
-        const total = sum(points);
-        return {
-          data: {
-            data: points,
-            label: `${playerLabel(player)} (${total})`,
-            backgroundColor: playerColor(player, false),
-            borderColor: "black",
-            borderWidth: 1,
-          },
-          points: total,
-        };
-      }),
-    (d) => -d.points
-  ).map((d) => d.data);
-
-  const tooltipOptions: DeepPartial<TooltipOptions<"bar">> = {
-    callbacks: {
-      afterTitle(this: TooltipModel<"bar">, items: TooltipItem<"bar">[]): string | string[] {
-        const dataIndex = items[0].dataIndex;
-        return ` (${victoryPointSources[dataIndex].description})`;
-      },
-    },
-  };
-
-  return {
-    type: "bar",
-    data: {
-      labels: victoryPointSources.map((s) => s.label),
-      datasets: datasets,
-    },
-    options: {
-      plugins: {
-        tooltip: tooltipOptions,
-        title: {
-          text: "Victory points of all players",
-          display: true,
-        },
-      } as any,
-      responsive: true,
-    },
-  };
-}
-
-export function newResourcesBarChart(data: Engine, canvas: HTMLCanvasElement): ChartConfiguration<"bar"> {
-  const datasets: ChartDataset<"bar">[] = chartPlayerOrder(data)
-    .filter((pl) => data.player(pl).faction != null)
-    .map((pl) => {
-      const points = resourceDetails(data, pl, canvas, resourceSources, "last").map((f) => f.getDataPoints()[0]);
-      const player = data.player(pl);
-      return {
-        data: points,
-        label: playerLabel(player),
-        backgroundColor: playerColor(player, false),
-        borderColor: "black",
-        borderWidth: 1,
-      };
-    });
-
-  return {
-    type: "bar",
-    data: {
-      labels: resourceSources.map((s) => s.label),
-      datasets: datasets,
-    },
-    options: {
-      plugins: {
-        title: {
-          text: "Resources of all players",
-          display: true,
-        },
-      },
-      responsive: true,
-    },
-  };
-}
-
-export function newVictoryPointsLineChart(
-  data: Engine,
-  canvas: HTMLCanvasElement,
-  player: PlayerEnum
-): ChartConfiguration<"line"> {
-  const numbers = player === PlayerEnum.All ? chartPlayerOrder(data) : [player];
-
-  let title: string;
-  let factories: DatasetFactory[];
-  if (numbers.length == 1) {
-    const p = numbers[0];
-    title = `Victory points of ${playerLabel(data.player(p))}`;
-    factories = victoryPointDetails(data, p, canvas, groupSources(victoryPointSources));
-  } else {
-    title = "Victory points of all players";
-    factories = numbers.map((p) =>
-      weightedSum(data, p, canvas, victoryPointDetails(data, p, canvas, victoryPointSources))
-    );
-  }
-
-  return lineChartConfig(
-    title,
-    canvas,
-    data,
-    factories.filter((f) => f != null),
-    player != PlayerEnum.All,
-    "all"
-  );
-}
-
-export function newResourcesLineChart(
-  data: Engine,
-  canvas: HTMLCanvasElement,
-  kind: PlayerEnum | ResourceKind
-): ChartConfiguration<"line"> {
-  let title: string;
-  let factories: DatasetFactory[];
-  if (typeof kind == "number") {
-    const numbers = kind === PlayerEnum.All ? chartPlayerOrder(data) : [kind as PlayerEnum];
-
-    if (numbers.length == 1) {
-      const player = numbers[0];
-      title = `Resources of ${playerLabel(data.player(player))}`;
-      factories = resourceDetails(data, player, canvas, resourceSources, "except-final");
-    } else {
-      const conversion = resourceSources
-        .filter((s) => s.weight > 0)
-        .map((s) => `${s.label}=${s.weight}`)
-        .join(", ");
-      title = `Resources of all players as if bought with power (${conversion})`;
-      factories = numbers.map((p) =>
-        weightedSum(data, p, canvas, resourceDetails(data, p, canvas, resourceSources, "except-final"))
-      );
-    }
-  } else {
-    const sources = resourceSources.filter((r) => r.type === (kind as Resource));
-
-    title = `${sources[0].plural} of all players`;
-    factories = chartPlayerOrder(data).map((p) => {
-      const pl = data.player(p);
-      if (!pl.faction) {
-        return null;
-      }
-
-      return {
-        backgroundColor: playerColor(pl, true),
-        label: playerLabel(pl),
-        fill: false,
-        getDataPoints: () => resourceDetails(data, p, canvas, sources, "except-final")[0].getDataPoints(),
-        weight: 1,
-      };
-    });
-  }
-
-  return lineChartConfig(
-    title,
-    canvas,
-    data,
-    factories.filter((f) => f != null),
-    false,
-    "except-final"
-  );
 }
